pyproject.toml
[project]
name = "ibelieveicanflypy"
version = "0.1.0"
description = ""
authors = [
    {name = "LucasTechAI",email = "lucas.mendestech@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy (==2.3.4)",
    "pandas (==2.3.3)",
    "scipy (==1.16.3)",
    "scikit-learn (==1.7.2)",
    "lightgbm (==4.6.0)",
    "matplotlib (==3.10.7)",
    "seaborn (==0.13.2)",
    "plotly (==6.5.0)",
    "statsmodels (==0.14.5)",
    "patsy (==1.0.2)",
    "python-dotenv (==1.2.1)",
    "joblib (==1.5.2)",
    "graphviz (==0.21)",
    "ipykernel (==7.1.0)",
    "ipython (==9.6.0)",
    "jupyter-client (==8.6.3)",
    "jupyter-core (==5.9.1)",
    "psutil (==7.1.3)",
    "black (==25.12.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
